requirements.txt
fastapi
uvicorn
yfinance
torch
torchvision
pandas
numpy
Pillow

 
